requirements.txt
numpy
scipy
pytorch
inflect
librosa
Unidecode
matplotlib
tensorboardX
